fix(ios): address second Copilot review round on badges wall
- BadgeShowcase.from now folds milestones from the same newest-per-key
  doc selection that already backed awardedAtByKey, instead of an
  arbitrary `first(where:)` match. A duplicate document for one badge
  key (denormalized badgeKey diverging from the doc id) could otherwise
  hand a milestone a stale name/date than the one awardedAtByKey reports
  for the same key. Added a regression test with the older duplicate
  listed first, which the old first-match code would have picked.
- BadgeText.subtitle now passes Int64 into String(format:), matching the
  %1$lld/%2$lld specifiers the generated catalog actually uses.
- Corrected FirebaseBadgesRepository's header comment: the earned-badges
  listener on users/{uid}/badges is PUBLIC read for any authenticated
  user per firebase/firestore.rules, not owner-only — only the progress
  callable is genuinely owner-scoped.

+++ apps/ios/KCC/Badges/BadgesWall.swift
@@ -345,7 +345,10 @@ private struct BadgeMedallion: View {
 enum BadgeText {
     /// `%1$lld of %2$lld unlocked`.
     static func subtitle(earned: Int, total: Int) -> String {
-        String(format: template("badgeShowcase.subtitle"), earned, total)
+        // The catalog string is `%1$lld of %2$lld unlocked` (a 64-bit format
+        // specifier); pass Int64 explicitly rather than relying on Int's
+        // CVarArg encoding to happen to match on every current ABI.
+        String(format: template("badgeShowcase.subtitle"), Int64(earned), Int64(total))
     }
 
     /// `Next: %1$@`.
